Add/Fix very basic flare rendering
Binds a new static entity ubo. Flare entity ubo should be updated when the projection matrix changes, but it works for basic rendering. Flares in cubemaps should be broken with this. Shader ubo is not bound properly, needs to be implemented. Need to get the shader data upfront when building the frame ubos. Definitly need to get rid of the glReadPixels to get the current depth value, as this creates a sync point.