Multiplayer RoffSystem Fixes (#1165)
* [MP] Fix ROFF system using incorrect time value.

The ROFF system previously used engine time instead of module time when setting the time values for trajectories of shared entities. This caused issues when the server module time has been reset on map change, because the game module would misinterpret the trajectories.

* [MP] Fix roff not working on 64 bit builds.

Change type of mVersion in TROFFHeader and TROFF2Header from long to int.