[SP] Rename MenuFontToReal to UI_MenuFontToReal and replace hardcoded font mapping with a dynamic list of up to 64 fonts.
This might help with possible corner cases where custom assets have the menu loads differents fonts in place of the vanilla ones. The maximum of 64 fonts was chosen, because I wanted to avoid dynamic allocation. As the vanilla game only loads 5 different fonts and as I am not aware of custom mods that add a lot more I chose the value 64 to hopefully leave enough room for mods.

The same change can NOT be applied to CG_MagicFontToReal, because the cgame module does not parse the menu files and does not know the order of fonts in menu files.