[JO] Fix level transition screen (#1254)
When redsaurus fixed the save screenshot (in ff210853d582adcf58578cba9ee2917607d49c7c ), they must have missed the double usage of SCR_GetScreenshot. Besides being used for the savegame thumbnails, it's used directly as an argument for DrawStrechRaw. DrawStrechRaw expects a RGBA buffer to draw to the screen. For drawing the level transition screen correctly I also had to flip the image in RE_GetScreenShot, so I also changed the flipping behaviour when saving the savegame.