[rend2] fixed crash without causing clumping
See issue #1189

The game will crash, if a map features weather (e.g. vjun1) and you noclip far enough beneath the map.

This was caused by an out-of-bounds exception caused by chunkIndex sometimes being negative.

The code change ensures the result is always in the range [0, 2] for each axis. This makes sure chunkIndex is always in [0, 8], which is safe for zoneOffsets[9].